requirements.txt
langchain
openai
faiss-cpu
unstructured
pytesseract
pdf2image
streamlit
python-dotenv
pandas
docx2txt
